refactor(db): Change amount columns to Numeric for financial precision

+++ backend/finmate/models/transaction_model.py
@@ -8,7 +8,7 @@ class Transactions(db.Model):
 
     id = db.Column(db.Integer, primary_key=True)
     transaction_type = db.Column(db.String(10), nullable=False) #income abo expense #potom na enum kak-to
-    amount = db.Column(db.Integer, nullable=False)
+    amount = db.Column(db.Numeric(10, 2), nullable=False)
     title = db.Column(db.String(128))
     note = db.Column(db.String(128), nullable=True)
     created_at = db.Column(db.DateTime(timezone=True), server_default=func.now(), nullable=False)
